refactor and add to do
