+++ bccsp/sw/keyderiv.go
@@ -23,11 +23,7 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/Hyperledger-TWGC/ccs-gm/sm2"
-
 	"github.com/hyperledger/fabric/bccsp"
-
-	bccspsm2 "github.com/hyperledger/fabric/bccsp/sm2"
 )
 
 type ecdsaPublicKeyKeyDeriver struct{}
@@ -154,98 +150,3 @@ func (kd *aesPrivateKeyKeyDeriver) KeyDeriv(k bccsp.Key, opts bccsp.KeyDerivOpts
 }
 
 // TWGC todo
-type sm2PublicKeyKeyDeriver struct{}
-
-func (kd *sm2PublicKeyKeyDeriver) KeyDeriv(k bccsp.Key, opts bccsp.KeyDerivOpts) (bccsp.Key, error) {
-	// Validate opts
-	if opts == nil {
-		return nil, errors.New("Invalid opts parameter. It must not be nil.")
-	}
-
-	sm2K := k.(*bccspsm2.SM2PublicKey)
-
-	switch opts.(type) {
-	// Re-randomized an SM2 public key
-	case *bccsp.SM2ReRandKeyOpts:
-		reRandOpts := opts.(*bccsp.SM2ReRandKeyOpts)
-		tempSK := &sm2.PublicKey{
-			Curve: sm2K.PubKey.Curve,
-			X:     new(big.Int),
-			Y:     new(big.Int),
-		}
-
-		var k = new(big.Int).SetBytes(reRandOpts.ExpansionValue())
-		var one = new(big.Int).SetInt64(1)
-		n := new(big.Int).Sub(sm2K.PubKey.Params().N, one)
-		k.Mod(k, n)
-		k.Add(k, one)
-
-		// Compute temporary public key
-		tempX, tempY := sm2K.PubKey.ScalarBaseMult(k.Bytes())
-		tempSK.X, tempSK.Y = tempSK.Add(
-			sm2K.PubKey.X, sm2K.PubKey.Y,
-			tempX, tempY,
-		)
-
-		// Verify temporary public key is a valid point on the reference curve
-		isOn := tempSK.Curve.IsOnCurve(tempSK.X, tempSK.Y)
-		if !isOn {
-			return nil, errors.New("Failed temporary public key IsOnCurve check.")
-		}
-		return &bccspsm2.SM2PublicKey{tempSK}, nil
-	default:
-		return nil, fmt.Errorf("Unsupported 'KeyDerivOpts' provided [%v]", opts)
-	}
-}
-
-// TWGC todo
-type sm2PrivateKeyKeyDeriver struct{}
-
-func (kd *sm2PrivateKeyKeyDeriver) KeyDeriv(k bccsp.Key, opts bccsp.KeyDerivOpts) (bccsp.Key, error) {
-	// Validate opts
-	if opts == nil {
-		return nil, errors.New("Invalid opts parameter. It must not be nil.")
-	}
-
-	sm2K := k.(*bccspsm2.SM2PrivateKey)
-
-	switch opts.(type) {
-	// Re-randomized an ECDSA private key
-	case *bccsp.SM2ReRandKeyOpts:
-		reRandOpts := opts.(*bccsp.SM2ReRandKeyOpts)
-		tempSK := &sm2.PrivateKey{
-			PublicKey: sm2.PublicKey{
-				Curve: sm2K.PrivKey.Curve,
-				X:     new(big.Int),
-				Y:     new(big.Int),
-			},
-			D: new(big.Int),
-		}
-
-		var k = new(big.Int).SetBytes(reRandOpts.ExpansionValue())
-		var one = new(big.Int).SetInt64(1)
-		n := new(big.Int).Sub(sm2K.PrivKey.Params().N, one)
-		k.Mod(k, n)
-		k.Add(k, one)
-
-		tempSK.D.Add(sm2K.PrivKey.D, k)
-		tempSK.D.Mod(tempSK.D, sm2K.PrivKey.PublicKey.Params().N)
-
-		// Compute temporary public key
-		tempX, tempY := sm2K.PrivKey.PublicKey.ScalarBaseMult(k.Bytes())
-		tempSK.PublicKey.X, tempSK.PublicKey.Y =
-			tempSK.PublicKey.Add(
-				sm2K.PrivKey.PublicKey.X, sm2K.PrivKey.PublicKey.Y,
-				tempX, tempY,
-			)
-
-		// Verify temporary public key is a valid point on the reference curve
-		isOn := tempSK.Curve.IsOnCurve(tempSK.PublicKey.X, tempSK.PublicKey.Y)
-		if !isOn {
-			return nil, errors.New("Failed temporary public key IsOnCurve check.")
-		}
-		return &bccspsm2.SM2PrivateKey{tempSK}, nil
-	default:
-		return nil, fmt.Errorf("Unsupported 'KeyDerivOpts' provided [%v]", opts)
-	}
-}
